Update setwgtw and fix Makefile
The new vector normalization in setwgtw was updated based on the
recommendations from Jerry Sellwood - this will ensure that the intended
exponentiation is performed. Thanks Jerry!

Additionally, issues introduced into the Makefile by the separation of
the compilation of 3rd-party libraries was fixed by disabling the
implicit rule for linking an executable from a .o file, requiring that
the linking pattern rule specified in the Makefile is used.