Update setwgtw to correct normalizing zero vector
A bug was introduced with the new warp vector normalization code, where
the zero vector would be normalized to (Infinity, Infinity). This would
result in the weight for that pixel being NaN, which propagated to
the main func calculation and gave a Chi^2 of NaN, causing issues in the
minimizer.

Now, we recognize when we are getting the weight for the center of the
galaxy, and set ct, st, rp2 = 0 so that the weight calculated will be 0.