Work around error in trig functions (#3)
* Work around error in trig functions

Certain compilers / libc's don't have bit-perfect trig functions, so we
now perform trig calculations in double-precision and convert back to
single-precision after performing the calculation.

Alternatively, one spot in setwgtw.f has been modified to perform an
alternative calculation: instead of using trigonometric functions, we
use trig. identities to compute sine and cosine directly from the
triangle's side lengths.

* Update setwgtw to correct normalizing zero vector

A bug was introduced with the new warp vector normalization code, where
the zero vector would be normalized to (Infinity, Infinity). This would
result in the weight for that pixel being NaN, which propagated to
the main func calculation and gave a Chi^2 of NaN, causing issues in the
minimizer.

Now, we recognize when we are getting the weight for the center of the
galaxy, and set ct, st, rp2 = 0 so that the weight calculated will be 0.

* Use smaller "tiny" value than the tiny constant