Bypass libquadmath (#7)
* Avoid linking with libquadmath

Libquadmath is licensed under the LGPL, which is not compatible with
this project. Therefore, we cannot statically link against it. This adds
a C file containing stubs for all functions called from libquadmath,
allowing us to get around the "dependency" on it.

* Fix libquadmath_stub

You can't use __attribute__((alias ...)) on macOS, so we just call the
lqm_panic function from the stubs.