Add support for compiling on macOS (Darwin) (#8)
* Add special compilation support for macOS

On macOS (Darwin), we can't just use "-static" to link the executable,
since it's impossible to statically link against Apple's libSystem.
Instead, we need to find all the static object files for all libraries
that we're using, and manually specify them on the command line so they
will be used instead of the dynamic versions. The newly added
DarwinFindLibraryArchive.sh script does this. Modifications were made to
the Makefile to invoke this script to find the libraries used and to do
a "semi-static" link where all libraries except libSystem are linked
statically.

Additionally, some helper shell scripts had to be modified to work with
non-GNU sed.

* Add -lSystem to link against macOS system library